Use transform_nest for set_field
